Improve logging conf
Add more job debugging

Sentry in settings.LOGGING

Switch raven transport to Requests (RQ):
Error delivery to Sentry is known to be unreliable with RQ when using async transports (the default is). So you are encouraged to use the HTTPTransport or RequestsHTTPTransport when creating your client.

Buuut, this'll probably make normal error reporting much slower (synchronous) so we'll want to see if we can have one setting for task runners and another for web workers.