Change member profile updates to use a fixed update form instead of dynamic.
Dynamic form was trying to be a bit too clever, whilst working on #328
the fields appeared out of order. With a fixed form this is a fixed order.
The ability to change required/optinal is still available within the model.

Partially reverts 78ba3be

Rename DynamicUpdateProfile and associated names to MemberProfileUpdate
in line with convention.

Update tests to reflect these changes.

Closes #328